Split up firebase api calls
