+++ src/components/Profile/Company/api.ts
@@ -1,0 +1,17 @@
+import { Firestore, doc, getDoc } from 'firebase/firestore';
+import { Production } from '../../Profile/Company/types';
+
+export const getProduction = async (
+  firebaseStore: Firestore,
+  productionId: string
+) => {
+  const docRef = doc(firebaseStore, 'productions', productionId);
+  const docSnap = await getDoc(docRef);
+
+  if (docSnap.exists()) {
+    const data = docSnap.data() as Production;
+    return data;
+  } else {
+    return false;
+  }
+};
